+++ app/jobs/stellar_base/application_job.rb
@@ -1,4 +1,7 @@
 module StellarBase
-  class ApplicationJob < ActiveJob::Base
+  class ApplicationJob
+
+    include Sidekiq::Worker
+
   end
 end