+++ app/models/stellar_base/account_subscription.rb
@@ -1,0 +1,4 @@
+module StellarBase
+  class AccountSubscription < ApplicationRecord
+  end
+end