+++ app/services/stellar_base/account_subscriptions/save_cursor.rb
@@ -1,0 +1,12 @@
+module StellarBase
+  class SaveCursor
+
+    extend LightService::Action
+    expects :account_subscription, :operations
+
+    executed do |c|
+      c.account_subscription.update_attributes!(cursor: c.operations.last.id)
+    end
+
+  end
+end