+++ lib/stellar_base/factories/account_subscription.rb
@@ -1,0 +1,8 @@
+FactoryBot.define do
+
+  factory :stellar_base_account_subscription, class: "StellarBase::AccountSubscription" do
+    address { Stellar::Account.random.address }
+    cursor { nil }
+  end
+
+end