+++ spec/spec_helper.rb
@@ -5,6 +5,7 @@ require "pry"
 require "vcr"
 require "rspec/rails"
 require "stellar-sdk"
+require "rspec-sidekiq"
 
 SPEC_DIR = Pathname.new(File.dirname(__FILE__))
 ROOT_DIR = SPEC_DIR.join("..")